+++ tests/utilities.py
@@ -7,7 +7,7 @@ from vvrest.vault import Vault
 from .settings import credentials_file, parameters_file
 
 
-def get_vault_object(user_web_token=None, jwt=None, auto_jwt=False, docapi_enabled=True):
+def get_vault_object(user_web_token=None, jwt=None, docapi_enabled=True):
     """
     :param user_web_token: string UUID(version=4), used for user impersonation
     :param jwt: str, JSON Web Token
@@ -19,10 +19,10 @@ def get_vault_object(user_web_token=None, jwt=None, auto_jwt=False, docapi_enabl
 
     if docapi_enabled:
         vault = Vault(credentials['url'], credentials['customer_alias'], credentials['database_alias'],
-                      credentials['client_id'], credentials['client_secret'], user_web_token, jwt, auto_jwt)
+                      credentials['client_id'], credentials['client_secret'], user_web_token, jwt)
     else:
         vault = Vault(credentials['docapi_not_enabled']['url'], credentials['docapi_not_enabled']['customer_alias'], credentials['docapi_not_enabled']['database_alias'],
-                      credentials['docapi_not_enabled']['client_id'], credentials['docapi_not_enabled']['client_secret'], user_web_token, jwt, auto_jwt)
+                      credentials['docapi_not_enabled']['client_id'], credentials['docapi_not_enabled']['client_secret'], user_web_token, jwt)
 
     return vault