+++ vvrest/vault.py
@@ -6,7 +6,7 @@ from .services.user_service import UserService
 
 
 class Vault:
-    def __init__(self, url, customer_alias, database_alias, client_id, client_secret, user_web_token=None, jwt=None, auto_jwt=False):
+    def __init__(self, url, customer_alias, database_alias, client_id, client_secret, user_web_token=None, jwt=None):
         """
         if user_web_token is passed in, then vv will authenticate on behalf of the user that
         the web_token belongs to. if user_web_token is not passed in (default=None), then
@@ -18,7 +18,6 @@ class Vault:
         :param client_secret: str, example: khN18YAZPe6F3Z0tc2W0HXCb487jm0wgwe6kNffUNf0=
         :param user_web_token: str UUID(version=4), passed in if authentication is user impersonation
         :param jwt: string, JSON Web Token
-        :param auto_jwt: bool, if True jwt will be fetched for user and used in auth headers
         """
         self.url = url
         self.customer_alias = customer_alias
@@ -30,22 +29,19 @@ class Vault:
         self.token = self.get_access_token()
         self.base_url = self.get_base_url()
         self.docapi_url = None
-        self.auto_jwt = auto_jwt
 
-        # get jwt for user + reset auth headers if is auto_jwt
-        if self.auto_jwt:
+        # get jwt for user if not provided
+        if not self.jwt:
             user_service = UserService(self)
             resp = user_service.get_user_jwt()
             self.jwt = resp['data']['token']
-            self.token = self.get_access_token()
 
         # if docapi is enabled set docapi_url
         config_service = ConfigService(self)
-        if self.jwt:
-            docapi_config = config_service.get_docapi_config()
-            if 'data' in docapi_config:
-                if docapi_config['data']['isEnabled']:
-                    self.docapi_url = docapi_config['data']['apiUrl']
+        docapi_config = config_service.get_docapi_config()
+        if 'data' in docapi_config:
+            if docapi_config['data']['isEnabled']:
+                self.docapi_url = docapi_config['data']['apiUrl']
 
     def get_access_token(self):
         """
@@ -96,5 +92,13 @@ class Vault:
         :return: dict
         """
         headers = {'Authorization': 'Bearer ' + self.token.access_token}
+
+        return headers
+    
+    def get_jwt_auth_headers(self):
+        """
+        :return: dict
+        """
+        headers = {'Authorization': 'Bearer ' + self.jwt}
 
         return headers