+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 # v2.0.0
 - add `docapi_document_service.DocApiDocumentService`
 - add `docapi_file_service.DocApiFileService`
-- add `auto_jwt` to `Vault`
+- add `docapi_search_service.DocApiSearchService`
+- add auto fetching `jwt` on `vault` instantiation for `docapi`
 
 # v1.8.0
 - add `delete_document_revision` to `DocumentService`